+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## **Version 2.0.0.1-RC1**
 
 - Fixed a bug with moving minimap icon.
+- Fixed disenchant table for classic TBC.
 
 ## **Version 2.0.0.0-RC1**